# submission/api/openapi.yaml
openapi: 3.1.0
info:
  title: LabelHub Core API
  version: "1.0.0"
  summary: HTTP edge of the LabelHub data-annotation platform.
  description: >
    Portable OpenAPI description of LabelHub's **HTTP route handlers**
    (`src/app/api/**`). These thin external edges authenticate, then delegate to
    the library layer. They cover the **machine-to-machine** surfaces:
    workspace-API-key ingest/read endpoints, admin (Supabase session) export
    endpoints, public operational endpoints, and the LLM gateway proxy.


    LabelHub deliberately keeps **internal mutations off HTTP** — the browser UI
    creates tasks, submits/reviews annotations, and runs payouts through
    **Next.js Server Actions** (`'use server'` functions), NOT these routes. That
    RPC surface is not expressible in OpenAPI; it is documented in the companion
    `API.md` ("Server-Action RPC surface"). This file documents only what is
    reachable over HTTP.


    ## Authentication models

    Three distinct models are in play (see `securitySchemes`):

    1. **Workspace API key** — machine-to-machine. A `lh_ws_*` (or rate-limited
       public `lh_demo_*`) bearer token. Only the SHA-256 hash is stored; the
       workspace is *inferred from the key*, so these endpoints take no
       `workspaceId`. Accepted on `Authorization: Bearer <key>`, bare
       `Authorization: <key>`, `x-api-key: <key>`, or `x-labelhub-api-key: <key>`.
    2. **Supabase user session** — cookie session from `@supabase/ssr`. Used by
...

...
servers:
...
paths:
...
  /api/proxy/{kind}/{path}:
    post:
      tags: [Gateway proxy]
      summary: LLM provider proxy with trajectory capture
      description: >
        Catch-all proxy. Forwards a chat-completions / messages request to the
        registered provider's upstream (`<connection.baseUrl>/<path>`), streams
        or buffers the response back, and captures the round-trip as a
        trajectory in the after()-window. Workspace API key auth; per-connection
        and per-key RPM limits; optional topic-scope system-prompt injection.
        `model` and a non-empty `messages` array are required in the body.
      operationId: proxyLlm
      security:
        - workspaceApiKey: []
      parameters:
        - name: kind
          in: path
          required: true
          description: Provider key (e.g. `openai`, `anthropic`, `doubao`, `deepseek`, `qwen`, `moonshot`).
          schema: { type: string }
        - name: path
          in: path
          required: true
          description: Upstream path forwarded verbatim (e.g. `chat/completions` or `v1/messages`).
          schema: { type: string }
      requestBody:
        required: true
        description: Provider-native chat request (OpenAI-compat or Anthropic Messages).
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model: { type: string }
                messages:
                  type: array
                  items: { type: object, additionalProperties: true }
                stream: { type: boolean }
              additionalProperties: true
      responses:
        "200":
          description: Upstream response (JSON or SSE stream), passed through.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
            text/event-stream:
              schema: { type: string }
        "400":
          description: Bad JSON / missing model / missing messages.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxyError" }
        "401":
          description: Missing/invalid API key.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxyError" }
        "404":
          description: Unknown provider (`UNKNOWN_PROVIDER`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxyError" }
        "429":
          description: Rate limited (`RATE_LIMITED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxyError" }
        "502":
          description: Upstream not configured / unreachable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxyError" }
        "500":
          description: Internal error.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxyError" }
    get:
      tags: [Gateway proxy]
      summary: Method-not-allowed hint
      operationId: proxyLlmGet
      parameters:
        - name: kind
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: path
          required: true
          schema: { type: string }
      responses:
        "405":
          description: Use POST.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxyError" }

  # ───────────────────────── Operational (public) ────────────────────────────
...
components:
  securitySchemes:
    workspaceApiKey:
      type: http
      scheme: bearer
      description: >
        Workspace API key (`lh_ws_*`, or rate-limited public `lh_demo_*`).
        Accepted on `Authorization: Bearer <key>`, bare `Authorization: <key>`,
        `x-api-key: <key>`, or `x-labelhub-api-key: <key>`. The workspace is
        inferred from the key.
    supabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: >
        Supabase cookie session (`@supabase/ssr`). The route resolves the target
        workspace, then enforces a workspace-scoped guard
        (`requireWorkspaceAdmin` / `requireWorkspaceMember`). Cookie name varies
        by Supabase project ref.
    adminToken:
      type: http
      scheme: bearer
      description: >
        Out-of-band ops secret from the `ADMIN_DIAG_TOKEN` env. Send via
        `Authorization: Bearer <token>` or `x-admin-token`; `?token=` query is a
        soft-deprecated fallback.
  schemas:
...
    ProxyError:
      type: object
      description: Error envelope used by the gateway proxy (adds a `type` tag).
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            code: { type: string }
            type: { type: string, const: labelhub_proxy }
      required: [error]
...